refactor(racingcar package): Move racingcar domain package
