+++ src/main/kotlin/step3/racingcar/domain/cars/Position.kt
@@ -1,4 +1,4 @@
-package step3.racingcar.domain
+package step3.racingcar.domain.cars
 
 data class Position(val value: Int) {
     init {
