+++ src/main/kotlin/step3/racingcar/domain/cars/Racing.kt
@@ -1,6 +1,7 @@
-package step3.racingcar.domain
+package step3.racingcar.domain.cars
 
-import step3.racingcar.domain.MovingStatus.MOVE_ONE_STEP_FORWARD
+import step3.racingcar.domain.status.MovingStatus.MOVE_ONE_STEP_FORWARD
+import step3.racingcar.domain.engine.Engine
 
 data class Racing(private val engine: Engine) {
     fun race(position: Position): Position {
