+++ src/main/kotlin/step3/racingcar/domain/engine/Engine.kt
@@ -1,4 +1,4 @@
-package step3.racingcar.domain
+package step3.racingcar.domain.engine
 
 fun interface Engine {
     fun generateCriterionValueToRace(): Int
